requirement.txt
absl-py==2.1.0
astunparse==1.6.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
dm-tree==0.1.8
docstring_parser==0.16
etils==1.9.2
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
googleapis-common-protos==1.63.2
grpcio==1.65.5
h5py==3.11.0
idna==3.7
immutabledict==4.2.0
importlib_resources==6.4.3
imutils==0.5.4
kagglehub==0.2.9
keras==3.5.0
keras-core==0.1.7
keras-cv==0.9.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
numpy==1.26.4
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.3
pillow==10.4.0
promise==2.3
protobuf==4.25.4
psutil==6.0.0
pyarrow==17.0.0
pybind11==2.13.6
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
regex==2024.7.24
requests==2.32.3
rich==13.7.1
setuptools==72.2.0
simple_parsing==0.1.5
six==1.16.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-datasets==4.9.6
tensorflow-metadata==1.15.0
termcolor==2.4.0
tflite-support==0.1.0a1
toml==0.10.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
Werkzeug==3.0.3
wheel==0.44.0
wrapt==1.16.0
zipp==3.20.0
